Sample the next polling interval from an exponential distribution.
Rather than ramp it up from a short interval up to 5 seconds.

I find this doesn't work so well. Most of the time it's fine, when the
server has something to send during the 0.0 interval that follows a
successful send or receive. But when the server takes slightly longer to
prepare something to send, the client then has to wait for the polling
interval to expire (mean 10 s) rather than having a quick interval of
100 ms.