Add proxy URL tests using IP addresses.
I assumed that you should be able to give a proxy host as a domain name,
but it turns out that proposal 232 doesn't actually say that
(https://trac.torproject.org/projects/tor/ticket/12125#comment:3). Add
some IP addresses in addition to host names so we have that covered as
well.