Honor proxy even with http URLs in UTLSRoundTripper.
Formerly, RoundTrip simply deferred to the global httpRoundTripper (the
one we would use if we were not using uTLS) for any http (as opposed to
https) requests. However, we allow setting a proxy on UTLSRoundTripper
that can be different from the proxy set on the global httpRoundTripper.

In practice, the global httpRoundTripper probably *would* have the same
proxy setting, just because of how the PT protocol only allows a single
proxy and how we initialize everything together. But there's nothing at
the UTLSRoundTripper layer that enforces that, so it's better if we
don't assume it. UTLSRoundTripper is already slightly weird because of
the rule about not mixing http/1.1 and h2; this change means that http
URLs don't add any additional weirdness.