Detect errors of proxy.onRequest.
It turns out that if an error occurs in proxy.onRequest, Firefox will
ignore it and continue on as if there were no proxy. This can happen,
for example, if the "type" of a ProxyInfo isn't one of the recognized
types, or if the ProxyInfo is missing a field like "host".
https://bugzilla.mozilla.org/show_bug.cgi?id=1528873
To prevent silent failures like this, we register another pair of event
listeners. Unlike the headers and proxy event listeners, these remain
static for all requests, and are not overwritten with each new request:
 * proxy.onError detects when an error occurs.
 * webRequest.onBeforeRequest cancels all requests after an error
I made the error condition be persistent because it's not something that
should arise during normal operation. proxy.onError only gets called
when an error occurs in proxy.onRequest, and that only happens when the
proxy specification is bogus. It doesn't get called when there's an
network error, for example.

In order to make exceptions in proxy.onRequest be noticed by
proxy.onError, I had to make it return a rejection promise rather than
throw an exception. An exception just gets logged to the browser console
and nothing else.