Use a non-empty ServerName in TestProxyHTTPSCONNECT.
The crypto/tls server, since go1.12, requires that if a server_name
extension is present, it is not empty. With go1.12 and later, this test
was failing with
	proxy_test.go:95: local error: tls: unexpected message

The specific commit that added the empty server_name check was:
https://go-review.googlesource.com/c/go/+/144115
https://github.com/golang/go/commit/4c8b09e9183390d6ab80d3f53a9fe5f6ace92f06#diff-f2f5faeac5da94dc7735456f798177b53421f827b8bfc242b56f116b8e5c7e62R264
Note the comment "enforcing a couple more 'must not be empty' on the
unmarshaling side".

Wireshark also complains in a packet capture:
	[Expert Info (Warning/Protocol): Vector length 0 is smaller than minimum 1]

https://bugs.torproject.org/tpo/anti-censorship/pluggable-transports/meek/40002